// src/module/invoice/usecase/find-invoice/find-invoice.dto.ts
export type FindInvoiceUsecaseInputDto = {
    id: string;
}

export type FindInvoiceUsecaseOutputDto = {
    id: string;
    name: string;
    document: string;
    address: {
      street: string;
      number: string;
      complement: string;
      city: string;
      state: string;
      zipCode: string;
    };
    items: {
      id: string;
      name: string;
      price: number;
    }[];
    total: number;
    createdAt: Date;
}